backend/requirements.txt
# Django Core
Django==4.2.11
djangorestframework==3.15.1
django-cors-headers==4.3.1
django-filter==24.1
django-environ==0.11.2

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Auth & Security
djangorestframework-simplejwt==5.3.1
django-otp==1.4.0
cryptography==42.0.5

# Cache & Queue
redis==5.0.3
django-redis==5.4.0
celery==5.3.6
django-celery-beat==2.6.0
django-celery-results==2.5.1
flower==2.0.1

# Storage
Pillow==10.2.0
django-storages==1.14.2
boto3==1.34.69  # S3-compatible object storage

# Performance
gunicorn==21.2.0
whitenoise==6.6.0
django-silk==5.1.0  # SQL profiling in dev

# API documentation
drf-spectacular==0.27.1

# Utils
python-slugify==8.0.4
phonenumbers==8.13.30
django-phonenumber-field==7.3.0
openpyxl==3.1.2  # Excel import
pandas==2.2.1    # Data processing

# Monitoring & Logging
sentry-sdk==1.43.0
django-structlog==8.0.0

# Testing
pytest==8.1.1
pytest-django==4.8.0
pytest-cov==5.0.0
factory-boy==3.3.0
faker==24.3.0

# Dev tools
black==24.3.0
isort==5.13.2
flake8==7.0.0
